#325: Created Endpoint for retrieving solutions

+++ src/main/java/org/benetech/mathshare/repository/ProblemSetRevisionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProblemSetRevisionRepository extends JpaRepository<ProblemSetRevision, Long> {
+public interface ProblemSetRevisionRepository extends JpaRepository<ProblemSetRevision, Integer> {
 
     ProblemSetRevision findOneByProblemSetAndReplacedBy(ProblemSet problemSet, ProblemSetRevision replacedBy);
