+++ src/main/java/org/benetech/mathshare/repository/ProblemSolutionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProblemSolutionRepository extends JpaRepository<ProblemSolution, Long> {
+public interface ProblemSolutionRepository extends JpaRepository<ProblemSolution, Integer> {
 
     ProblemSolution findOneByEditCode(long editCode);
 }
