+++ src/main/java/org/benetech/mathshare/repository/SolutionRevisionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SolutionRevisionRepository extends JpaRepository<SolutionRevision, Long> {
+public interface SolutionRevisionRepository extends JpaRepository<SolutionRevision, Integer> {
 
     SolutionRevision findOneByShareCode(long shareCode);
